Move `FieldSelections` to `helpers.ts`
