+++ graphql-utils/src/index.ts
@@ -1,3 +1,5 @@
 export { hasFields } from "./has-fields";
-export { resolveSelections, FieldSelections } from "./resolve-selections";
+export { fieldMapToDot, FieldSelections } from "./helpers";
+export { resolveFieldMap } from "./resolve-field-map";
 export { resolveFields } from "./resolve-fields";
+export { resolveSelections } from "./resolve-selections";
